feat(ARC-3792): convert fe search url to be search api url via admin-core-config
https://meemoo.atlassian.net/browse/ARC-3792
